+++ lib/flop/nimble_schemas.ex
@@ -109,6 +109,7 @@ defmodule Flop.NimbleSchemas do
               type: {:tuple, [:atom, :atom, :keyword_list]},
               required: true
             ],
+            orderer: [type: {:tuple, [:atom, :atom, :keyword_list]}],
             ecto_type: [type: :any],
             bindings: [type: {:list, :atom}],
             operators: [type: {:list, :atom}]