+++ lib/flop/validation.ex
@@ -359,8 +359,12 @@ defmodule Flop.Validation do
 
   defp unsupported_cursor_field?(struct, field) do
     case Flop.Schema.field_info(struct, field) do
-      %FieldInfo{extra: %{type: type}} when type in [:compound, :alias] -> true
-      _ -> false
+      %FieldInfo{extra: %{type: type}}
+      when type in [:compound, :alias, :custom] ->
+        true
+
+      _ ->
+        false
     end
   end