+++ test/support/fruit.ex
@@ -3,6 +3,7 @@ defmodule MyApp.Fruit do
   Defines an Ecto schema for testing.
   """
   use Ecto.Schema
+  import Ecto.Query
 
   alias MyApp.Owner
 
@@ -17,7 +18,7 @@ defmodule MyApp.Fruit do
              :owner_attributes,
              :owner_extra
            ],
-           sortable: [:id, :name],
+           sortable: [:id, :name, :name_custom],
            join_fields: [
              owner_attributes: [
                binding: :owner,
@@ -30,6 +31,13 @@ defmodule MyApp.Fruit do
                field: :extra,
                path: [:owner, :extra],
                ecto_type: :map
+             ]
+           ],
+           custom_fields: [
+             name_custom: [
+               filter: {__MODULE__, :name_custom_filter, []},
+               orderer: {__MODULE__, :name_custom_orderer, []},
+               ecto_type: :string
              ]
            ],
            default_limit: 60,
@@ -48,5 +56,14 @@ defmodule MyApp.Fruit do
     field :references, {:array, :binary_id}
 
     belongs_to :owner, Owner
+  end
+
+  def name_custom_filter(query, %Flop.Filter{value: value}, _opts) do
+    where(query, [f], f.name == ^value)
+  end
+
+  def name_custom_orderer(query, direction, _opts) do
+    send(self(), {:orderer, direction})
+    order_by(query, [f], [{^direction, f.name}])
   end
 end